# spec/openapi.yaml
openapi: 3.0.3
info:
  title: Family Meal Planner API (V1)
  description: |
    Next.js Route Handlers under `/api/*`. All week keys are **Sunday** calendar dates in **America/Toronto** (`YYYY-MM-DD`), matching `lib/week-boundary.ts`.

    **Id semantics:** `PATCH` / `PATCH …/have` use **ShoppingLineGroup** ids. `DELETE /api/shopping-list/{id}` uses **ShoppingLineContribution** ids.

    Canonical narrative: `api-routes.md`, `openapi.yaml` (this file), and `api-sequences.md` should stay aligned when routes change.
  version: "1.0.0"
servers:
  - url: /
    description: Next.js app origin (e.g. http://localhost:3000)
...
paths:
  /api/meal-plan:
    get:
      tags: [meal-plan]
      summary: Load weekly plan
      description: Returns nested `WeeklyPlan` or 404 if no row exists for the resolved week.
      parameters:
        - $ref: "#/components/parameters/WeekStartQuery"
      responses:
        "200":
          description: Prisma-shaped weekly plan (days ordered Sun→Sat, slots breakfast→dinner).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyPlan"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          description: No plan for this week
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: "No plan found for this week"
    post:
      tags: [meal-plan]
      summary: Save full weekly plan (last write wins)
      description: Validates body, egg rule, then upserts plan/days/slots in a transaction. Recomputes `proteinWarning` server-side.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealPlanSaveRequest"
      responses:
        "200":
          description: Reloaded full plan after save
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyPlan"
        "400":
          description: Validation or egg rule violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  parameters:
    WeekStartQuery:
      name: weekStart
      in: query
      required: false
      schema:
        type: string
        pattern: "^\\d{4}-\\d{2}-\\d{2}$"
      description: Sunday `YYYY-MM-DD` in America/Toronto. Omit → current week start.
  responses:
    BadRequestError:
      description: Invalid query or JSON body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    ServerError:
      description: Unexpected server/database error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
        field:
          type: string
          description: Optional field path for validation errors
    DayOfWeek:
      type: string
      enum:
        - SUNDAY
        - MONDAY
        - TUESDAY
        - WEDNESDAY
        - THURSDAY
        - FRIDAY
        - SATURDAY
    MealSlotType:
      type: string
      enum: [BREAKFAST, LUNCH, DINNER]
...
    MealSlot:
      type: object
      required:
        - id
        - dayPlanId
        - slot
        - mainMealText
        - proteinWarning
        - isQuick
        - isMakeAhead
        - isEasy
        - needsTime
        - toddlerFriendly
      properties:
        id:
          type: string
        dayPlanId:
          type: string
        slot:
          $ref: "#/components/schemas/MealSlotType"
        mainMealText:
          type: string
        proteinWarning:
          type: boolean
        isQuick:
          type: boolean
        isMakeAhead:
          type: boolean
        isEasy:
          type: boolean
        needsTime:
          type: boolean
        toddlerFriendly:
          type: boolean
        toddlerNote:
          type: string
          nullable: true
    MealSlotInput:
      type: object
      required:
        - slot
        - mainMealText
        - isQuick
        - isMakeAhead
        - isEasy
        - needsTime
        - toddlerFriendly
      properties:
        slot:
          $ref: "#/components/schemas/MealSlotType"
        mainMealText:
          type: string
        isQuick:
          type: boolean
        isMakeAhead:
          type: boolean
        isEasy:
          type: boolean
        needsTime:
          type: boolean
        toddlerFriendly:
          type: boolean
        toddlerNote:
          type: string
          nullable: true
    DayPlan:
      type: object
      required:
        - id
        - weeklyPlanId
        - dayOfWeek
        - isTrip
        - mealSlots
      properties:
        id:
          type: string
        weeklyPlanId:
          type: string
        dayOfWeek:
          $ref: "#/components/schemas/DayOfWeek"
        isTrip:
          type: boolean
        tripNotes:
          type: string
          nullable: true
        mealSlots:
          type: array
          items:
            $ref: "#/components/schemas/MealSlot"
    DayPlanInput:
      type: object
      required:
        - dayOfWeek
        - isTrip
        - mealSlots
      properties:
        dayOfWeek:
          $ref: "#/components/schemas/DayOfWeek"
        isTrip:
          type: boolean
        tripNotes:
          type: string
          nullable: true
        mealSlots:
          type: array
          items:
            $ref: "#/components/schemas/MealSlotInput"
    WeeklyPlan:
      type: object
      required:
        - id
        - weekStartSunday
        - createdAt
        - updatedAt
        - days
      properties:
        id:
          type: string
        weekStartSunday:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        days:
          type: array
          items:
            $ref: "#/components/schemas/DayPlan"
    MealPlanSaveRequest:
      type: object
      required:
        - weekStartSunday
        - days
      properties:
        weekStartSunday:
          type: string
          description: Sunday date `YYYY-MM-DD` or ISO instant matching Toronto Sunday midnight (see route parser).
        days:
          type: array
          items:
            $ref: "#/components/schemas/DayPlanInput"
...